requirements.txt
flask
mysql.connector
hashlib
threading
datetime
pyjwt
gevent-websocket
Flask-SocketIO
eventlet